refactor: split parser and formatter into per-concern files
parser.go retains core infrastructure (Parse, advance, expect helpers,
parseIdentList, parseStatement). DDL parsing moves to parse_ddl.go, SELECT
parsing to parse_select.go, and the raw expression reader to parse_expr.go.

formatter.go retains shared infrastructure (Format, kw, indent,
writeCommaList, formatStatement). DDL formatting moves to format_ddl.go and
SELECT formatting to format_select.go.

Test files mirror the source split: parser_test.go renamed to
parse_select_test.go; formatter_test.go trimmed to golden infrastructure;
format_ddl_test.go and format_select_test.go added for unit tests.

Closes #74
